+++ packages/solidity-devops/js/utils/chain.js
@@ -79,7 +79,8 @@ const readBlockscoutOptions = (chainName) => {
 }
 
 const readSourcifyOptions = (chainName) => {
-  return '--verifier sourcify'
+  const url = tryReadEnv(chainName, 'VERIFIER_URL')
+  return '--verifier sourcify' + (url ? ` --verifier-url ${url}` : '')
 }
 
 /**