+++ packages/synapse-interface/constants/routes.ts
@@ -1,5 +1,4 @@
 import {
-  EXPLORER_PATH,
   SWAP_PATH,
   STAKE_PATH,
   POOLS_PATH,
@@ -8,7 +7,6 @@ import {
   BRIDGE_PATH,
   SOLANA_BRIDGE_LINK,
   SYN_TOKEN_LINK,
-  STAKE_SYN_FOR_CX_URL,
 } from './urls'
 
 export interface RouteObject {
@@ -49,11 +47,6 @@ export const NAVIGATION: RouteObject = {
       startsWith: STAKE_PATH,
     },
   },
-  Analytics: {
-    path: EXPLORER_PATH,
-    text: 'Explorer',
-    match: null,
-  },
   SYN: {
     path: SYN_TOKEN_LINK,
     text: '$SYN',
@@ -62,11 +55,6 @@ export const NAVIGATION: RouteObject = {
   Solana: {
     path: SOLANA_BRIDGE_LINK,
     text: 'Solana Bridge',
-    match: null,
-  },
-  Cortex: {
-    path: STAKE_SYN_FOR_CX_URL,
-    text: 'Stake SYN for CX',
     match: null,
   },
 }