+++ docs/bridge/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.15](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.6.14...@synapsecns/bridge-docs@0.6.15) (2026-03-05)
+
+**Note:** Version bump only for package @synapsecns/bridge-docs
+
+
+
+
+
+## [0.6.14](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.6.13...@synapsecns/bridge-docs@0.6.14) (2026-02-13)
+
+**Note:** Version bump only for package @synapsecns/bridge-docs
+
+
+
+
+
+## [0.6.13](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.6.12...@synapsecns/bridge-docs@0.6.13) (2025-11-26)
+
+**Note:** Version bump only for package @synapsecns/bridge-docs
+
+
+
+
+
+## [0.6.12](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.6.11...@synapsecns/bridge-docs@0.6.12) (2025-11-24)
+
+**Note:** Version bump only for package @synapsecns/bridge-docs
+
+
+
+
+
+## [0.6.11](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.6.10...@synapsecns/bridge-docs@0.6.11) (2025-09-09)
+
+**Note:** Version bump only for package @synapsecns/bridge-docs
+
+
+
+
+
+## [0.6.10](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.6.9...@synapsecns/bridge-docs@0.6.10) (2025-08-18)
+
+**Note:** Version bump only for package @synapsecns/bridge-docs
+
+
+
+
+
 ## [0.6.9](https://github.com/synapsecns/sanguine/compare/@synapsecns/bridge-docs@0.6.8...@synapsecns/bridge-docs@0.6.9) (2025-04-10)
 
 **Note:** Version bump only for package @synapsecns/bridge-docs