+++ packages/contracts-adapter/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-adapter@1.0.3...@synapsecns/contracts-adapter@1.0.4) (2025-09-26)
+
+**Note:** Version bump only for package @synapsecns/contracts-adapter
+
+
+
+
+
 ## [1.0.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/contracts-adapter@1.0.2...@synapsecns/contracts-adapter@1.0.3) (2025-06-30)
 
 **Note:** Version bump only for package @synapsecns/contracts-adapter