+++ packages/rest-api/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.26](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.25...@synapsecns/rest-api@1.13.26) (2025-12-02)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.13.25](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.24...@synapsecns/rest-api@1.13.25) (2025-11-27)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.13.24](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.23...@synapsecns/rest-api@1.13.24) (2025-11-26)
+
+
+### Bug Fixes
+
+* replace RPCs that are not responding ([#3846](https://github.com/synapsecns/sanguine/issues/3846)) ([1c67df0](https://github.com/synapsecns/sanguine/commit/1c67df040dec29cf2a16cd427226c60ef7f1bd2b))
+
+
+
+
+
+## [1.13.23](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.22...@synapsecns/rest-api@1.13.23) (2025-11-13)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.13.22](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.21...@synapsecns/rest-api@1.13.22) (2025-11-07)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.13.21](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.20...@synapsecns/rest-api@1.13.21) (2025-10-27)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.13.20](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.19...@synapsecns/rest-api@1.13.20) (2025-09-19)
+
+
+### Bug Fixes
+
+* **rest-api:** error when no routes are found in bridgeV2 ([#3805](https://github.com/synapsecns/sanguine/issues/3805)) ([37dc6c3](https://github.com/synapsecns/sanguine/commit/37dc6c39d8f3d10b797f23d106dcdbe710ecfc57))
+
+
+
+
+
+## [1.13.19](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.18...@synapsecns/rest-api@1.13.19) (2025-09-19)
+
+
+### Bug Fixes
+
+* **rest-api:** error when no routes found ([#3804](https://github.com/synapsecns/sanguine/issues/3804)) ([965088f](https://github.com/synapsecns/sanguine/commit/965088f7d625cef346db66d21de85a32a1ee631b))
+
+
+
+
+
 ## [1.13.18](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.13.17...@synapsecns/rest-api@1.13.18) (2025-07-22)
 
 **Note:** Version bump only for package @synapsecns/rest-api